requirements.txt
pytelegrambotapi==4.27.0
requests==2.32.4
python-dotenv==1.1.1